Added a load_csv option. Made value.name_numeric to float, since it was rounding floats.

+++ src/dqt_api/models.py
@@ -44,7 +44,7 @@ class Value(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(50))
-    name_numeric = db.Column(db.Numeric)
+    name_numeric = db.Column(db.Float)
     description = db.Column(db.String(500))
 
     def __init__(self, name, description=None):
